doc(README.md): fix "add to path" command
VyOS's `vbash` interprets command lines starting with `sh` as "show"
commands. We need to either give a full path (e.g., `/bin/sh`), or
a more explicit command name.

Given that (as far as I'm aware) the VyOS post-run scripts run under
`vbash`, I figure that should be a fine choice of shell to use for this.